Extract generation preview submit hook
